+++ src/main/java/com/chalchal/chalchalserver/auth/controller/AuthController.java
@@ -3,12 +3,11 @@ package com.chalchal.chalchalserver.auth.controller;
 import com.chalchal.chalchalserver.auth.dto.LoginUserRequest;
 import com.chalchal.chalchalserver.auth.dto.UserRequest;
 import com.chalchal.chalchalserver.auth.dto.UserResponse;
-import com.chalchal.chalchalserver.domain.auth.dto.*;
 import com.chalchal.chalchalserver.auth.service.UserService;
-import com.chalchal.chalchalserver.auth.service.UserTokenInfoService;
 import com.chalchal.chalchalserver.auth.jwt.JwtUtils;
 import com.chalchal.chalchalserver.auth.domain.User;
 import com.chalchal.chalchalserver.auth.domain.UserTokenInfo;
+import com.chalchal.chalchalserver.auth.service.UserTokenInfoService;
 import com.chalchal.chalchalserver.global.dto.ErrorResponse;
 import com.chalchal.chalchalserver.global.dto.ResultResponse;
 import com.chalchal.chalchalserver.global.exception.ErrorCode;
