+++ src/main/java/com/chalchal/chalchalserver/todo/controller/TodoController.java
@@ -3,8 +3,8 @@ package com.chalchal.chalchalserver.todo.controller;
 import com.chalchal.chalchalserver.auth.domain.User;
 import com.chalchal.chalchalserver.todo.dto.TodoListSaveRequest;
 import com.chalchal.chalchalserver.todo.dto.TodoListUpdateRequest;
-import com.chalchal.chalchalserver.todo.service.TodoService;
 import com.chalchal.chalchalserver.global.dto.ResultResponse;
+import com.chalchal.chalchalserver.todo.service.TodoService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
